+++ src/hammerjs/index.ts
@@ -14,10 +14,6 @@ export {TapRecognizer as Tap} from './recognizers/tap';
 export {PanRecognizer as Pan} from './recognizers/pan';
 export {SwipeRecognizer as Swipe} from './recognizers/swipe';
 export {PinchRecognizer as Pinch} from './recognizers/pinch';
-export {
-  TwoFingerPanRecognizer as TwoFingerPan,
-  TwoFingerPinchRecognizer as TwoFingerPinch
-} from './recognizers/two-finger';
 export {RotateRecognizer as Rotate} from './recognizers/rotate';
 export {PressRecognizer as Press} from './recognizers/press';
 
@@ -27,5 +23,6 @@ export type {TapRecognizerOptions} from './recognizers/tap';
 export type {PanRecognizerOptions} from './recognizers/pan';
 export type {SwipeRecognizerOptions} from './recognizers/swipe';
 export type {PinchRecognizerOptions} from './recognizers/pinch';
+export type {InputCoherenceCondition} from './recognizers/attribute';
 export type {RotateRecognizerOptions} from './recognizers/rotate';
 export type {PressRecognizerOptions} from './recognizers/press';